Update repository sync to synchronize all repos everytime

+++ src/main/webapp/app/shared/components/list/list.component.ts
@@ -327,8 +327,10 @@ export class ListComponent<T> implements OnInit, OnDestroy {
    * @param dropEvent The drag & drop event
    */
   public onDropEvent(dropEvent: CdkDragDrop<string[]>) {
-    moveItemInArray(this.objectsPaged.content, dropEvent.previousIndex, dropEvent.currentIndex);
-    this.drop();
+    if (dropEvent.previousIndex != dropEvent.currentIndex) {
+      moveItemInArray(this.objectsPaged.content, dropEvent.previousIndex, dropEvent.currentIndex);
+      this.drop();
+    }
   }
 
   /**
